Build 51

+++ RetroChess/Clock.cpp
@@ -1,6 +1,5 @@
 #include "Clock.h"
 
-
 Clock::Clock()
 {
 	time = 0;
@@ -17,7 +16,7 @@ void Clock::Start()
 
 void Clock::Stop()
 {
-	time += std::chrono::duration_cast<std::chrono::duration<int,std::milli>>(std::chrono::high_resolution_clock::now() - begin).count();
+	time += std::chrono::duration_cast<std::chrono::duration<unsigned long long,std::milli>>(std::chrono::high_resolution_clock::now() - begin).count();
     begin = std::chrono::high_resolution_clock::now();
 }
