+++ RetroChess/Clock.h
@@ -8,7 +8,8 @@ public:
 	std::chrono::high_resolution_clock::time_point begin;
     //std::chrono::high_resolution_clock::time_point end;
 
-	unsigned long time;
+	unsigned long long time;
+	//std::chrono::time_point<std::chrono::high_resolution_clock> t;
 
 	Clock();
 	~Clock();
